Added empty solution.
At the outset, I'll be writing this project without tests, as an example
of a code base that may otherwise be nice and well-structured, but is
difficult to test.

Later, I will refactor (or perhaps, more precicely, strangle) the code by
adding tests to a new code structure which will replace the old structure.