Introduced ApplicantProcessor,
which can be used to produce renderings for an Applicant instance. While
it doesn't implement any interface, it's a pure function, so it can be
used in Facade Tests to verify composed behaviour.

However, the method signature suggest that a possible refactoring could
make IMortgageApplicationProcessor a generic type like this:

public interface IProcessor<T>
{
    IEnumerable<IRendering> ProduceRenderings(T item);
}